# Avoid specializing all of ForwardDiff on every equation